# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A powerful URL shortening service with analytics, QR codes, and link management.",
        "title": "Patrn.ink URL Shortener API",
        "termsOfService": "https://patrn.ink/terms",
        "contact": {
            "name": "API Support",
            "url": "https://patrn.ink/support",
            "email": "support@patrn.ink"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/preview": {
            "get": {
                "description": "Fetches and returns metadata (title, description, favicon) for a given URL",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Preview"
                ],
                "summary": "Get URL preview",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL to fetch preview for",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.LinkPreview"
                        }
                    },
                    "400": {
                        "description": "Invalid or missing URL",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch URL",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/{code}/preview": {
            "get": {
                "description": "Returns URL preview metadata for an existing short link",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Preview"
                ],
                "summary": "Get link preview by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.LinkPreview"
                        }
                    },
                    "404": {
                        "description": "Link not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch preview",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/{code}/qr": {
            "get": {
                "description": "Generates a QR code PNG image for the short URL",
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "QR Code"
                ],
                "summary": "Generate QR code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PNG image",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "URL not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "410": {
                        "description": "Link has expired",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.LinkPreview": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "favicon": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "description": "API token for programmatic access",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT Bearer token or API token (prefix with \"Bearer \")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}